Even more spaghetti

+++ Form1.cs
@@ -28,6 +28,11 @@ namespace BSPlaylistDownloader
 
         private string BSPath;
         private string CustomLevelsPath;
+        private string PlaylistsPath;
+
+        private bool playlistsloaded = false;
+        private bool autostart = true;
+
         public Form1()
         {
             InitializeComponent();
@@ -43,8 +48,10 @@ namespace BSPlaylistDownloader
             else
                 textBox1.Text = location;
 
+            checkBox1.Checked = Convert.ToBoolean(ConfigurationManager.AppSettings.Get("AutoStart"));
+
             var appVersion = Assembly.GetExecutingAssembly().GetName().Version;
-            label3.Text = String.Format("{0} (pre-alpha)", appVersion);
+            label3.Text = String.Format("{0}", appVersion);
         }
 
         private void button1_Click(object sender, EventArgs e)
@@ -96,20 +103,22 @@ namespace BSPlaylistDownloader
                 config.Save(ConfigurationSaveMode.Modified);
                 ConfigurationManager.RefreshSection("appSettings");
 
-                if (!Directory.Exists(location + "\\Playlists"))
+                BSPath = location;
+                PlaylistsPath = location + "\\Playlists\\";
+                CustomLevelsPath = location + "\\Beat Saber_Data\\CustomLevels\\";
+
+                if (!Directory.Exists(PlaylistsPath))
                 {
                     richTextBox1.SelectionColor = Color.DarkRed;
                     richTextBox1.AppendText("Directory \"Playlists\" does not exists.\n");
-                    Directory.CreateDirectory(location + "/Playlists");
-                    if (Directory.Exists(location + "\\Playlists"))
+                    Directory.CreateDirectory(PlaylistsPath);
+                    if (Directory.Exists(PlaylistsPath))
                     {
                         richTextBox1.SelectionColor = Color.Green;
                         richTextBox1.AppendText("Created \"Playlists\" directory.\n");
                     }
                     richTextBox1.SelectionColor = richTextBox1.ForeColor;
                 }
-
-                CustomLevelsPath = location + "\\Beat Saber_Data\\CustomLevels\\";
 
                 if (!Directory.Exists(CustomLevelsPath))
                 {
@@ -123,8 +132,6 @@ namespace BSPlaylistDownloader
                     }
                     richTextBox1.SelectionColor = richTextBox1.ForeColor;
                 }
-
-                BSPath = location;
 
                 button2.Enabled = true;
             }
@@ -140,22 +147,23 @@ namespace BSPlaylistDownloader
             openFileDialog1.Multiselect = true;
             if (openFileDialog1.ShowDialog() == DialogResult.OK)
             {
-                var watch = System.Diagnostics.Stopwatch.StartNew();
-
-                button2.Enabled = false;
-
-                richTextBox1.SelectionColor = Color.Green;
-                richTextBox1.AppendText(String.Format("Loaded {0} playlists \n", openFileDialog1.FileNames.Length));
-                richTextBox1.SelectionColor = richTextBox1.ForeColor;
+                playlistsloaded = true;
                 playlists = openFileDialog1.FileNames;
+                richTextBox1.SelectionColor = Color.Green;
+                richTextBox1.AppendText(String.Format("[I] Loaded {0} playlists \n", openFileDialog1.FileNames.Length));
+                richTextBox1.SelectionColor = richTextBox1.ForeColor;
 
-                await RunParsePlaylistsAsync();
+                progressBar1.Visible = false;
+                progressBar1.Value = 0;
+                label5.Text = "";
 
-                button2.Enabled = true;
-
-                watch.Stop();
-                var elapsedMs = watch.ElapsedMilliseconds;
-                richTextBox1.AppendText(String.Format("\n\nTotal execution time: {0}ms \n", elapsedMs));
+                if (autostart)
+                {
+                    await RunParsePlaylistsAsync();
+                } else
+                {
+                    button3.Enabled = true;
+                }
             }
         }
 
@@ -167,6 +175,10 @@ namespace BSPlaylistDownloader
 
         private async Task RunParsePlaylistsAsync()
         {
+            var watch = System.Diagnostics.Stopwatch.StartNew();
+
+            button2.Enabled = false;
+            button3.Enabled = false;
             int playlistscount = playlists.Length;
             progressBar1.Visible = true;
             progressBar1.Maximum = playlistscount;
@@ -199,22 +211,35 @@ namespace BSPlaylistDownloader
             {
                 richTextBox1.AppendText(value);
             });
+            var richbox1color = new Progress<Color>(value =>
+            {
+                richTextBox1.SelectionColor = value;
+            });
 
-            await Task.Run(() => ParsePlaylists(progress1, label5text, progress2, progress2max, label6text, richbox1text));
+            await Task.Run(() => ParsePlaylists(progress1, label5text, progress2, progress2max, label6text, richbox1text, richbox1color));
 
             progressBar1.Value = progressBar1.Maximum;
             label5.Text = String.Format("Completed");
             label6.Text = "";
             label6.Visible = false;
             progressBar2.Visible = false;
+            button2.Enabled = true;
+            playlistsloaded = false;
+            playlists = null;
+            watch.Stop();
+            var elapsedMs = watch.ElapsedMilliseconds;
+            richTextBox1.AppendText(String.Format("\n\n[I] Total execution time: {0}ms \n", elapsedMs));
         }
 
-        private async Task ParsePlaylists(IProgress<int> progress1, IProgress<string> label5text, IProgress<int> progress2, IProgress<int> progress2max, IProgress<string> label6text, IProgress<string> richbox1text)
+        private async Task ParsePlaylists(IProgress<int> progress1, IProgress<string> label5text, IProgress<int> progress2, IProgress<int> progress2max, IProgress<string> label6text, IProgress<string> richbox1text, IProgress<Color> richbox1color)
         {
             int i = 0;
             var regex = new Regex(@"\r\n?|\n|\t", RegexOptions.Compiled);
             foreach (string file in playlists)
             {
+                string filename = Path.GetFileName(file);
+                string destfilepath = PlaylistsPath + filename;
+
                 ++i;
                 string jsonData = File.ReadAllText(file);
                 Playlist playlist = JsonConvert.DeserializeObject<Playlist>(jsonData);
@@ -223,7 +248,17 @@ namespace BSPlaylistDownloader
                 string author = playlist.playlistAuthor;
                 int songcount = playlist.songs.Count;
 
-                richbox1text.Report(String.Format("Playlist \"{0}\" by {1} loaded, {2} songs. \n", title, author, songcount));
+                if (File.Exists(destfilepath))
+                {
+                    richbox1color.Report(Color.Orange);
+                    richbox1text.Report(String.Format("[W]Playlist {0} already exists, skipping.\n", title));
+                    richbox1color.Report(richTextBox1.ForeColor);
+                    continue;
+                }
+
+                richbox1color.Report(Color.Green);
+                richbox1text.Report(String.Format("[I] Loaded {0} songs from playlist {1} by {2}.\n", songcount, title, author));
+                richbox1color.Report(richTextBox1.ForeColor);
 
                 label5text.Report(String.Format("{0} {1}", title, i));
 
@@ -236,54 +271,89 @@ namespace BSPlaylistDownloader
                     label6text.Report(String.Format("{0} ({1}/{2})", songtitle, ii, songcount));
                     progress2.Report(ii);
                     string songhash = playlistitem.hash;
-                    await Task.Run(() => DownloadSong(richbox1text, songhash));
+                    await Task.Run(() => DownloadSong(richbox1text, richbox1color, songhash));
                 }
 
                 Thread.Sleep(250);
 
-                if (progress1 != null)
-                    progress1.Report(i);
+                progress1.Report(i);
             }
         }
 
-        private void DownloadSong(IProgress<string> richbox1text, string hash)
+        private void DownloadSong(IProgress<string> richbox1text, IProgress<Color> richbox1color, string hash)
         {
-            using (WebClient wc = new WebClient())
+            try
             {
-                var jsonData = wc.DownloadString("https://api.beatsaver.com/maps/hash/" + hash);
-
-                //var song = JsonConvert.DeserializeObject<JToken>(jsonData);
-
-                dynamic convertObj = JObject.Parse(jsonData);
-
-                string songID = convertObj.id;
-                string songName = convertObj.metadata.songName;
-                string levelAuthorName = convertObj.metadata.levelAuthorName;
-                levelAuthorName = levelAuthorName.Replace(":", "");
-
-                string foldername = String.Format("{0} ({1} - {2})", songID, songName, levelAuthorName);
-                string folderpath = CustomLevelsPath + foldername;
-
-                if (!Directory.Exists(folderpath))
+                using (WebClient wc = new WebClient())
                 {
-                    string downloadUrl = convertObj.versions[0].downloadURL;
-                    string filepath = String.Format(folderpath + ".zip");
-                    try
+                    var jsonData = wc.DownloadString("https://api.beatsaver.com/maps/hash/" + hash);
+
+                    //var song = JsonConvert.DeserializeObject<JToken>(jsonData);
+
+                    dynamic convertObj = JObject.Parse(jsonData);
+
+                    string songID = convertObj.id;
+                    string songName = convertObj.metadata.songName;
+                    string levelAuthorName = convertObj.metadata.levelAuthorName;
+                    levelAuthorName = levelAuthorName.Replace(":", "");
+
+                    string foldername = String.Format("{0} ({1} - {2})", songID, songName, levelAuthorName);
+                    string folderpath = CustomLevelsPath + foldername;
+
+                    if (!Directory.Exists(folderpath))
                     {
-                        wc.DownloadFile(downloadUrl, filepath);
-                        ZipFile.ExtractToDirectory(filepath, folderpath);
-                        File.Delete(filepath);
+                        string downloadUrl = convertObj.versions[0].downloadURL;
+                        string filepath = String.Format(folderpath + ".zip");
+                        try
+                        {
+                            if (File.Exists(filepath))
+                                File.Delete(filepath);
+                            wc.DownloadFile(downloadUrl, filepath);
+                            ZipFile.ExtractToDirectory(filepath, folderpath);
+                            File.Delete(filepath);
+                            richbox1color.Report(Color.Green);
+                            richbox1text.Report(String.Format("[I]\tDownloaded {0}\n", foldername));
+                            richbox1color.Report(richTextBox1.ForeColor);
+                        }
+                        catch (Exception e)
+                        {
+                            richbox1color.Report(Color.DarkRed);
+                            richbox1text.Report(String.Format("[E]\tError while downloading {0} from {1}\n", foldername, downloadUrl));
+                            richbox1color.Report(richTextBox1.ForeColor);
+                        }
                     }
-                    catch (Exception e)
+                    else
                     {
-                        richbox1text.Report(String.Format("{0}\n{1}\n", foldername, downloadUrl));
+                        richbox1color.Report(Color.Orange);
+                        richbox1text.Report(String.Format("[W]\t{0} already exists, skipping.\n", foldername));
+                        richbox1color.Report(richTextBox1.ForeColor);
                     }
-                    richbox1text.Report(String.Format("Downloaded {0} \n", foldername));
                 }
+            }
+            catch (Exception e)
+            {
+                richbox1text.Report(String.Format("[E]\tError while fetching data\n"));
             }
         }
 
         private string Trunc(string s, int len) => s?.Length > len ? String.Format("{0}...", s.Substring(0, len)) : s;
+
+        private void checkBox1_CheckedChanged(object sender, EventArgs e)
+        {
+            autostart = !autostart;
+
+            Configuration config = ConfigurationManager.OpenExeConfiguration(Application.ExecutablePath);
+            config.AppSettings.Settings.Remove("AutoStart");
+            config.AppSettings.Settings.Add("AutoStart", Convert.ToString(autostart));
+            config.Save(ConfigurationSaveMode.Modified);
+            ConfigurationManager.RefreshSection("appSettings");
+        }
+
+        private async void button3_Click(object sender, EventArgs e)
+        {
+            if(playlistsloaded)
+                await RunParsePlaylistsAsync();
+        }
     }
 
     public class Playlist
